requirements.txt
pandas
numpy
scikit-learn
nltk
scipy
beautifulsoup4
requests
lxml
streamlit
flask
python-dotenv
